+++ bin/assets_cleaner.dart
@@ -1,6 +1,7 @@
 import 'package:args/command_runner.dart';
 import 'package:assets_cleaner/commands/clean_command.dart';
 import 'package:assets_cleaner/commands/init_command.dart';
+import 'package:assets_cleaner/commands/trash_command.dart';
 import 'package:assets_cleaner/commands/unused_command.dart';
 
 void main(List<String> args) {
@@ -12,5 +13,6 @@ void main(List<String> args) {
     ..addCommand(UnusedCommand())
     ..addCommand(CleanCommand())
     ..addCommand(InitCommand())
+    ..addCommand(TrashCommand())
     ..run(args);
 }